Refactoring.
Hiding some internal handler implementation details in the plugin package.  And renaming some classes to be a little more consistent.
