+++ proxy-wasm-java-host/src/main/java/io/roastedroot/proxywasm/plugin/HttpCallResponseHandler.java
@@ -1,0 +1,5 @@
+package io.roastedroot.proxywasm.plugin;
+
+public interface HttpCallResponseHandler {
+    void call(HttpCallResponse response);
+}
