+++ proxy-wasm-jaxrs-quarkus-example/src/main/java/io/roastedroot/proxywasm/jaxrs/example/Resources.java
@@ -1,6 +1,6 @@
 package io.roastedroot.proxywasm.jaxrs.example;
 
-import io.roastedroot.proxywasm.jaxrs.NamedWasmPlugin;
+import io.roastedroot.proxywasm.jaxrs.WasmPlugin;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 
@@ -9,7 +9,7 @@ public class Resources {
 
     @Path("/ffiTests/reverse")
     @POST
-    @NamedWasmPlugin("ffiTests") // filter with ffiTests wasm plugin
+    @WasmPlugin("ffiTests") // filter with ffiTests wasm plugin
     public String ffiTests(String body) {
         return body;
     }
